% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\docType{class}
\name{GeneAnnotations-class}
\alias{GeneAnnotations-class}
\title{GeneAnnotations class}
\description{
An S4 class to represent a set of gene annotations built from
tall-skinny table with the columns "Gene" and "Term". This class provides
the functionality to compute several useful metrics based on the gene
annotations: distance matrix, term frequency of annotation,
functional similarity and enrichment.
}
\section{Slots}{

\describe{
\item{\code{raw_annotations}}{A data.frame having columns "Gene" and "Term" that
contains all annotations.}

\item{\code{name}}{The annotations name}

\item{\code{gene2term}}{A data.frame that stores the pivoted annotations across genes}

\item{\code{term2gene}}{A data.frame that stores the pivoted annotations across terms}

\item{\code{max_term_annotations}}{The maximum frequency of annotations, that is the maximum
number of terms associated to any gene}

\item{\code{func_similarity_methods}}{Stores the supported similarity methods}
}}

