% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{GeneAnnotations}
\alias{GeneAnnotations}
\title{GeneAnnotations constructor}
\usage{
GeneAnnotations(...)
}
\arguments{
\item{raw_annotations}{A data.frame having columns "Gene" and "Term" that
contains all annotations.}

\item{name}{The annotations name}
}
\value{
The GeneAnnotations object
}
\description{
The constructor for GeneAnnotations
}
\examples{
uniKeys <- AnnotationDbi::keys(org.Hs.eg.db::org.Hs.eg.db, keytype="SYMBOL")
cols <- c("PATH")
kegg_raw <- AnnotationDbi::select(org.Hs.eg.db::org.Hs.eg.db, keys=uniKeys, columns=cols, keytype="SYMBOL")
kegg_raw <- kegg_raw[, c(1, 2)]
colnames(kegg_raw) <- c("Gene", "Term")
kegg <- GeneAnnotations(raw_annotations = kegg_raw, name="KEGG-Human")
}

