% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_enrichment.R
\docType{class}
\name{GeneListEnrichment-class}
\alias{GeneListEnrichment-class}
\title{GeneListEnrichment class}
\description{
An S4 class to represent gene list enrichment results. This class depends on
a specific set of gene annotations represented by the class "GeneAnnotations".
The enrichment calculated is based on the Fisher exact test and we give
support to the usual multiple test adjustment methods.
}
\section{Slots}{

\describe{
\item{\code{gene_list}}{The list of genes over which the enrichment is computed.}

\item{\code{gene_annotations}}{The object GeneAnnotations to which the enrichment
refers.}

\item{\code{raw_enrichment}}{A data.frame with the enrichment results}
}}

