% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_enrichment.R
\name{GeneListEnrichment}
\alias{GeneListEnrichment}
\title{GeneListEnrichment constructor}
\usage{
GeneListEnrichment(...)
}
\arguments{
\item{gene_list}{The list of genes over which the enrichment is computed.}

\item{gene_annotations}{The object GeneAnnotations to which the enrichment
refers.}
}
\value{
The GeneAnnotations object
}
\description{
The constructor for GeneListEnrichment
}
\examples{
TODO
}

