% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{get_bc_similarity}
\alias{get_bc_similarity}
\title{get_bc_similarity()}
\usage{
get_bc_similarity(x, term1, term2)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}

\item{term1}{The first term on which to calculate the functional similarity.}

\item{term2}{The second term on which to calculate the functional similarity.}
}
\value{
The Bray-Curtis functional similarity between term1 and term2
}
\description{
Function that gets the Bray-Curtis functional similarity between two terms. That is
a measure of the shared terms in the annotations resources.
}
\examples{
kegg <- TCGAome::load_kegg()
random_kegg_term1 = kegg@term2gene$Term[runif(1, max = length(kegg@term2gene$Term))]
random_kegg_term2 = kegg@term2gene$Term[runif(1, max = length(kegg@term2gene$Term))]
get_bc_similarity(kegg, random_kegg_term1, random_kegg_term2)
}

