% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{get_enrichment}
\alias{get_enrichment}
\title{get_enrichment()}
\usage{
get_enrichment(x, gene_list)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}

\item{gene_list}{The list of genes on which to calculate the enrichment.}
}
\value{
The enrichment results stored in an object of class
GeneListEnrichment
}
\description{
Function that returns an object of class GeneListEnrichment with the
enrichment results on a given gene list on this gene annotations.
}
\examples{
kegg <- TCGAome::load_kegg()
gene_list <- c("ZNF638", "HNRNPU", "PPIAL4G", "RAPH1", "USP7", "SUMO1P3",
"TMEM189.UBE2V1", "ZNF837", "LPCAT4", "ZFPL1", "STAT3", "XRCC1",
"STMN1", "PGR", "RB1", "KDR", "YBX1", "YAP1", "FOXO3", "SYK", "RAB17",
"TTC8", "SLC22A5", "C3orf18", "ANKRA2", "LBR", "B3GNT5", "ANP32E",
"JOSD1", "ZNF695", "ESR1", "INPP4B", "PDK1", "TSC2", "AR", "HSPA1A",
"CDH3", "SMAD4", "CASP7", "GMPS", "NDC80", "EZH2", "MELK", "CDC45",
"CRY2", "KLHDC1", "MEIS3P1", "FBXL5", "EHD2", "CCNB1", "GSK3A",
"DVL3", "NFKB1", "COL6A1", "CCND1", "BAK1")
get_enrichment(kegg, gene_list)
}

