% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goa_enrichment.r
\name{get_enrichment_old}
\alias{get_enrichment_old}
\title{Computes the enrichment of a gene list into a given annotation resource (GOA and HPO are supported so far).
The enrichment is computed with the Fisher's exact test, multiple test
correction is optional.}
\usage{
get_enrichment_old(gene_list, annotation, pvalue_threshold = 0.01,
  adjustment_method = "none")
}
\arguments{
\item{gene_list}{The list of genes (entrez ids) for which we want to calculate the enrichment}

\item{annotation}{One of ["goa", "hpo"]}

\item{pvalue_threshold}{The threshold that determines significant results [default: 0.01]}

\item{adjustment_method}{Any method supported by p.adjust() [default: none]}
}
\description{
Computes the enrichment of a gene list into a given annotation resource (GOA and HPO are supported so far).
The enrichment is computed with the Fisher's exact test, multiple test
correction is optional.
}
\examples{
size = get_enrichment(c("57763", "81611", "367", "84002", "578", "51161", "840", "891",
"595", "8318", "1001", "1291", "1408", "1857", "30846", "2099",
"2146", "26234", "2309", "8833", "2931", "3192", "3303", "8821",
"9929", "3791", "122773", "3930", "254531", "9833", "10403",
"4790", "5163", "5241", "644591", "64284", "65059", "5925",
"6584", "4089", "6774", "3925", "100500808", "6850", "7249",
"123016", "7874", "7515", "10413", "4904", "7542", "27332",
"57116", "116412"), "goa")
}
\keyword{TCGAome}

