% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{get_functional_similarity}
\alias{get_functional_similarity}
\title{get_functional_similarity()}
\usage{
get_functional_similarity(x, term1, term2, distance_measure)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}

\item{term1}{The first term on which to calculate the functional similarity.}

\item{term2}{The second term on which to calculate the functional similarity.}

\item{distance_measure}{The binary distance method (one of UI, binary,
bray-curtis, cosine)}
}
\value{
The functional similarity between term1 and term2
}
\description{
Function that gets the functional similarity between two terms. That is
a measure of the shared terms in the annotations resources. This functional
similarity is implemented as a distance between binary vectors.
}
\examples{
kegg <- TCGAome::load_kegg()
random_kegg_term1 = kegg@term2gene$Term[runif(1, max = length(kegg@term2gene$Term))]
random_kegg_term2 = kegg@term2gene$Term[runif(1, max = length(kegg@term2gene$Term))]
get_functional_similarity(kegg, random_kegg_term1, random_kegg_term2, "cosine")
get_functional_similarity(kegg, random_kegg_term1, random_kegg_term2, "binary")
get_functional_similarity(kegg, random_kegg_term1, random_kegg_term2, "UI")
get_functional_similarity(kegg, random_kegg_term1, random_kegg_term2, "bray-curtis")
}

