% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{get_TCGAome_tumos_types}
\alias{get_TCGAome_tumos_types}
\title{Returns the supported tumor types.}
\usage{
get_TCGAome_tumos_types()
}
\description{
Returns the supported tumor types.
}
\examples{
get_TCGAome_tumos_types()
}
\keyword{TCGAome}

