% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{get_term_freq}
\alias{get_term_freq}
\title{get_term_freq()}
\usage{
get_term_freq(x, term)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}

\item{term}{The term on which we want the annotation frequency.}
}
\value{
The term's annotations frequency
}
\description{
Function that gets the frequency of annotation of a specific term. That is
the number of genes this term is associated to divided by the maximum term
annotation.
}
\examples{
kegg <- TCGAome::load_kegg()
random_kegg_term = kegg@term2gene$Term[runif(1, max = length(kegg@term2gene$Term))]
get_term_freq(kegg, random_kegg_term)
}

