% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_enrichment.R
\name{get_terms_clustering}
\alias{get_terms_clustering}
\title{get_terms_clustering()}
\usage{
get_terms_clustering(x, gene_annotations, distance_measure,
  significance_threshold, adj_method, max_clusters)
}
\arguments{
\item{x}{The GeneListEnrichment class on which the method will run.}

\item{gene_annotations}{The GeneAnnotations on which the enrichment was computed.}

\item{distance_measure}{The similarity distance metric employed.}

\item{significance_threshold}{The significance enrichment threshold}

\item{adj_method}{The multiple test correction method (any of those supported
by p.adjust())}
}
\value{
An object of class TermsClustering
}
\description{
Function that creates a new object of class TermsClustering. The clustering
is performed on a significant set of enrichment results that is defined by
the parameters passed to this function.
}
\examples{
TODO
}

