% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{load_kegg}
\alias{load_kegg}
\title{Reads KEGG pathways for Human genes and for all Uniprot.
No electronically inferred annotations (IEA)}
\usage{
load_kegg()
}
\description{
Reads KEGG pathways for Human genes and for all Uniprot.
No electronically inferred annotations (IEA)
}
\examples{
kegg = load_kegg()
}
\keyword{TCGAome}

