% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{load_omim}
\alias{load_omim}
\title{Reads OMIM diseases for Human genes.}
\usage{
load_omim()
}
\description{
Reads OMIM diseases for Human genes.
}
\examples{
omim = load_omim()
}
\keyword{TCGAome}

