% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{plot}
\alias{plot}
\title{plot()}
\usage{
plot(x)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}
}
\value{
The ggplot2 plot
}
\description{
Plots the gene annotations
}
\examples{
TCGAome::plot(hpo)
}

