% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_clustering.R
\name{plot_explained_variance}
\alias{plot_explained_variance}
\title{plot_explained_variance()}
\usage{
plot_explained_variance(x)
}
\arguments{
\item{x}{The GeneTermClustering class on which the method will run.}
}
\value{
The ggplot2 plot
}
\description{
Plots the explained variance of each MDS component
}
\examples{
TCGAome::plot_explained_variance(hpo_term_clustering)
}

