% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_clustering.R
\name{plot_mds}
\alias{plot_mds}
\title{plot_mds()}
\usage{
plot_mds(x, ..., all = FALSE)
}
\arguments{
\item{x}{The GeneTermClustering class on which the method will run.}

\item{all}{A flag indicating if showing all members of each cluster or
just the cluster representatives (default: FALSE)}
}
\value{
The ggplot2 plot
}
\description{
Plots the enriched terms on the three first components of the MDS results.
}
\examples{
TCGAome::plot_mds(hpo_term_clustering)
}

