% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter()}
\usage{
plot_scatter(x)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}
}
\value{
The ggplot2 plot
}
\description{
Plots the gene annotations scatter plot
}
\examples{
TCGAome::plot_scatter(hpo)
}

