% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_clustering.R
\name{plot_silhouette_analysis}
\alias{plot_silhouette_analysis}
\title{plot_silhouette_analysis()}
\usage{
plot_silhouette_analysis(x)
}
\arguments{
\item{x}{The GeneTermClustering class on which the method will run.}
}
\value{
The ggplot2 plot
}
\description{
Plots the silhouette average width of each number of clusters
}
\examples{
TCGAome::plot_silhouette_analysis(hpo_term_clustering)
}

