% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{plot_violin}
\alias{plot_violin}
\title{plot_violin()}
\usage{
plot_violin(x)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}
}
\value{
The ggplot2 plot
}
\description{
Plots the gene annotations violin distribution
}
\examples{
TCGAome::plot_violin(hpo)
}

