% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_clustering.R
\docType{class}
\name{TermsClustering-class}
\alias{TermsClustering-class}
\title{TermsClustering class}
\description{
An S4 class to represent the dimensionality reduction computed on
enrichment results. Given an object of class "GeneListEnrichment" and a
distance measure this class computes:
(1) the Partitioning Around Medoids (PAM) clustering
(2) the cluster representative selection based on the enrichment significance
and the frequency of annotation of the term to give priority to the most
specific terms
(3) the MultiDimensional Scaling (MDS) results on the cluster representatives
}
\section{Slots}{

\describe{
\item{\code{gene_list_enrichment}}{The object of class GeneListEnrichment that
that contains the enrichment results}

\item{\code{significance_threshold}}{The significance threshold applied to the enrichment
results. The significance threshold is a fixed value, if you want to modify
it you would need to create another object.}

\item{\code{adj_method}}{The multiple test adjustment method. Methods supported are
those supported by p.adjust()}

\item{\code{significant_results}}{A data.frame with the selected annotations terms
after applying the multiple test correction and the significance threshold.}

\item{\code{distance_measure}}{The distance measure employed for the clustering}

\item{\code{distance_matrix}}{The distance matrix between annotation terms by using
the given distance_measure}
}}

