% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_clustering.R
\name{TermsClustering}
\alias{TermsClustering}
\title{TermsClustering constructor}
\usage{
TermsClustering(...)
}
\arguments{
\item{gene_list_enrichment}{The object of class GeneListEnrichment that
that contains the enrichment results}

\item{distance_measure}{The distance measure employed for the clustering}

\item{significance_threshold}{The significance threshold applied to the enrichment
results. The significance threshold is a fixed value, if you want to modify
it you would need to create another object.}

\item{adj_method}{The multiple test adjustment method. Methods supported are
those supported by p.adjust()}
}
\value{
The TermsClustering object
}
\description{
The constructor for TermsClustering
}
\examples{
TODO
}

