% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_enrichment.R
\name{get_significant_results}
\alias{get_significant_results}
\title{get_significant_results()}
\usage{
get_significant_results(x, significance_thr, adj_method)
}
\arguments{
\item{x}{The GeneListEnrichment class on which the method will run.}

\item{significance_thr}{The significance enrichment threshold}

\item{adj_method}{The multiple test correction method (any of those supported
by p.adjust())}
}
\value{
A data.frame with the significant results
}
\description{
Function that gets the significant results from the enrichment results.
}
\examples{
TODO
}

