% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotations.R
\name{get_term_distance_matrix}
\alias{get_term_distance_matrix}
\title{get_term_distance_matrix()}
\usage{
get_term_distance_matrix(x, distance_measure, ..., terms_subset = NULL)
}
\arguments{
\item{x}{The GeneAnnotations class on which the method will run.}

\item{distance_measure}{The binary distance method (one of UI, binary,
bray-curtis, cosine)}

\item{terms_subset}{A subset of terms on which to calculate the distance matrix}
}
\value{
The distance matrix for the subset if provided or for all the terms
otherwise
}
\description{
Function that returns the distance matrix between all the terms.
}
\examples{
kegg <- TCGAome::load_kegg()
get_term_distance_matrix(kegg, "cosine")
}

