% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.r
\name{run.TCGAome}
\alias{run.TCGAome}
\title{Runs the TCGAome analysis pipeline on the given tumor types.}
\usage{
run.TCGAome(tumor_types)
}
\arguments{
\item{Vector}{of tumor types to be analyzed from those available at TCGA (run RTCGAToolbox::getFirehoseDatasets() to see all available types)}
}
\description{
Runs the TCGAome analysis pipeline on the given tumor types.
}
\examples{
run.TCGAome(c("BRCA", "OV"))
}
\keyword{TCGAome}

